requirements.txt
numpy==1.22.4
opencv_python==4.2.0.34
Pillow==9.4.0
streamlit==1.18.1
tensorflow==2.11.1
tensorflow_intel==2.11.0
